+++ libsummer/summer-download-youtube.c
@@ -74,7 +74,7 @@ create_youtube_url (SummerDownloadYoutube *self)
 }
 
 static void
-on_download_chunk (SummerWebBackend *web_backend, gint received, gint length, gpointer user_data)
+on_download_chunk (SummerWebBackend *web_backend, guint64 received, guint64 length, gpointer user_data)
 {
 	g_return_if_fail (SUMMER_IS_DOWNLOAD_YOUTUBE (user_data));
 	SummerDownload *self = SUMMER_DOWNLOAD (user_data);
